separate CDeviceResources class files
